chore: remove fiat from coin-framework currencies
