+++ libs/coin-framework/package.json
@@ -96,7 +96,6 @@
     "invariant": "^2.2.2",
     "lodash": "^4.17.21",
     "lru-cache": "^7.14.1",
-    "numeral": "^2.0.6",
     "prando": "^6.0.1",
     "rxjs": "^7.8.1"
   },
