+++ libs/coin-framework/src/currencies/index.ts
@@ -8,9 +8,9 @@ import {
 export { encodeURIScheme, decodeURIScheme } from "./CurrencyURIScheme";
 export { sanitizeValueString } from "./sanitizeValueString";
 export { parseCurrencyUnit } from "./parseCurrencyUnit";
-export { formatCurrencyUnit, formatCurrencyUnitFragment } from "./formatCurrencyUnit";
+export { formatCurrencyUnit, type formatCurrencyUnitOptions } from "./formatCurrencyUnit";
 export { toLocaleString } from "./BigNumberToLocaleString";
-export * from "./support";
+export { isCurrencySupported, listSupportedCurrencies, setSupportedCurrencies } from "./support";
 
 export const findCurrencyByTicker = (ticker: string): Currency | null | undefined =>
   findCryptoCurrencyByTicker(ticker) ||
